backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.3,<4.5
Pillow>=10.4,<11
redis>=5.0,<6
PyYAML>=6.0,<7

# Deployment
Dj-Database-URL>=2.2,<3
psycopg2-binary>=2.9,<3
gunicorn>=22,<23
whitenoise>=6.7,<7

# Optional (API docs)
drf-yasg>=1.21,<2
uritemplate>=4.1,<5
